pyproject.toml
[project]
name = "claude-web"
version = "0.1.0"
description = "Web-based Claude terminal service"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "nicegui>=1.4.0",
    "python-socketio>=5.10.0",
    "asyncssh>=2.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
